Add MessageFormat.parseResource(); rename resolveMessage() and resolve()
This change provides separate API surfaces for formatting a single message and
for formatting a resource of messages. Now, the `new MessageFormat()`
constructor builds a formatter for a single message, while the static method
`Messageformat.parseResource()` parses a resource of messages into a `Map` of
MessageFormat instances with a corresponding structure.

When making this change, it also makes sense to rename the `resolveMessage()`
method as just `resolve()`, given that the parent instance on which it's
operating is now a single message.