Add MessageFormat.parseResource() (#14)
This change provides separate API surfaces for formatting a single message and
for formatting a resource of messages. Now, the `new MessageFormat()`
constructor builds a formatter for a single message, while the static method
`Messageformat.parseResource()` parses a resource of messages into a Map of
MessageFormat instances with a corresponding structure.